added thrift client logging interceptor made logging interceptors conditional on property
