+++ starter/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,5 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-ru.trylogic.spring.boot.thrift.ThriftAutoConfiguration,\
-info.developerblog.spring.thrift.client.PoolConfiguration,\
+ru.trylogic.spring.boot.thrift.ThriftControllerAutoConfiguration,\
+info.developerblog.spring.thrift.client.ThriftClientAutoConfiguration,\
 info.developerblog.spring.thrift.client.ThriftClientBeanPostProcessor,\
 info.developerblog.spring.thrift.client.ThriftClientsMapBeanPostProcessor
